requirements.txt
pandas~=2.2.3
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
langid~=1.1.6
pytz~=2024.2
langdetect~=1.0.9
Pyrogram~=2.0.106
style~=1.1.0
pip~=24.2
filelock~=3.16.1
pika~=1.3.2
rabbit~=1.2.0
boto3~=1.35.90
botocore~=1.35.90
Telethon~=1.37.0
telegram~=0.0.1
APScheduler~=3.10.4
django-googlesearch~=0.2